+++ CHANGELOG.md
@@ -15,6 +15,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/) 
 ### Fixed
 
 - The `update` command no longer accidentally installs all tools when there are no tools to update
+- Made checking allowed filenames for binaries more robust
 
 ## [5.1.0] - 2026-05-25