+++ CHANGELOG.md
@@ -23,6 +23,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/) 
 - Made checking allowed filenames for binaries more robust
 - Default configuration now skips entries that are not supported on the current platform instead of producing an error
 - Create configuration file with the correct permissions directly
+- Fix null check in cache implementation
 
 ## [5.1.0] - 2026-05-25